+++ .eslintrc.js
@@ -1,11 +1,27 @@
 module.exports = {
+  env: {
+    browser: true,
+    es2021: true,
+  },
   extends: [
-    require.resolve('@contentful/eslint-config-extension'),
-    require.resolve('@contentful/eslint-config-extension/react.js'),
-    require.resolve('@contentful/eslint-config-extension/typescript.js'),
+    'eslint:recommended',
+    'plugin:react/recommended',
+    'plugin:react-hooks/recommended',
+    'plugin:@typescript-eslint/recommended',
   ],
-  plugins: ['eslint-plugin-import-helpers'],
+  parser: '@typescript-eslint/parser',
+  parserOptions: {
+    ecmaFeatures: {
+      jsx: true,
+    },
+    ecmaVersion: 'latest',
+    sourceType: 'module',
+  },
+  plugins: ['react', 'react-hooks', '@typescript-eslint', 'eslint-plugin-import-helpers'],
   rules: {
+    'react-hooks/exhaustive-deps': 'error',
+    '@typescript-eslint/no-explicit-any': 'warn',
+    '@typescript-eslint/no-unused-vars': ['error', { argsIgnorePattern: '^_' }],
     'import-helpers/order-imports': [
       'warn',
       {
@@ -15,5 +31,9 @@ module.exports = {
       },
     ],
     'no-restricted-imports': ['warn'],
+    'react/react-in-jsx-scope': 'off',
+  },
+  settings: {
+    react: { version: 'detect' },
   },
 };