+++ packages/rich-text/src/plugins/Hyperlink/components/EntityHyperlink.tsx
@@ -1,17 +1,15 @@
 import * as React from 'react';
 
-import { FieldAppSDK, Link } from '@contentful/app-sdk';
+import { Link } from '@contentful/app-sdk';
 import { Text } from '@contentful/f36-components';
 
-import { useContentfulEditor } from '../../../ContentfulEditorProvider';
-import { findNodePath, isChildPath } from '../../../internal/queries';
 import { Element, RenderElementProps } from '../../../internal/types';
-import { useSdkContext } from '../../../SdkProvider';
 import { useLinkTracking } from '../../links-tracking';
 import { useEntityInfo } from '../useEntityInfo';
 import { handleEditLink, handleRemoveLink } from './linkHandlers';
 import { LinkPopover } from './LinkPopover';
 import { styles } from './styles';
+import { useHyperlinkCommon } from './useHyperlinkCommon';
 
 export type HyperlinkElementProps = {
   element: Element & {
@@ -32,13 +30,9 @@ export type HyperlinkElementProps = {
 };
 
 export function EntityHyperlink(props: HyperlinkElementProps) {
-  const editor = useContentfulEditor();
-  const sdk: FieldAppSDK = useSdkContext();
-  const focus = editor.selection?.focus;
-  const { target } = props.element.data;
+  const { editor, sdk, isLinkFocused, pathToElement } = useHyperlinkCommon(props.element);
   const { onEntityFetchComplete } = useLinkTracking();
-  const pathToElement = findNodePath(editor, props.element);
-  const isLinkFocused = pathToElement && focus && isChildPath(focus.path, pathToElement);
+  const { target } = props.element.data;
 
   const tooltipContent = useEntityInfo({
     target,